A few trivial source cleanups while bringing up environment on new dev machine.
git-svn-id: https://svn.broadinstitute.org/gaag/pilon/trunk@4433 333ed61c-d966-4eaf-8115-f610d1735cec

+++ src/main/scala/org/broadinstitute/pilon/Utils.scala
@@ -19,7 +19,7 @@
 package org.broadinstitute.pilon
 
 object Utils {
-  def roundDiv(n: Long, d: Long) = if (d > 0) (n + d/2) / d else 0
+  def roundDiv(n: Long, d: Long) = if (d > 0) (n + d/2) / d else 0.toLong
   def roundDiv(n: Int, d: Int) = if (d > 0) (n + d/2) / d else 0
   def pct(n: Long, d: Long) = roundDiv(100 * n, d)
   def pct(n: Int, d: Int) = roundDiv(100 * n, d)
